Fix gzip support and add Default Application
