+++ product_code/authentication_service/impl/src/main/resources/application.conf
@@ -1,9 +1,6 @@
 #
 #
-play {
-    application.loader = de.upb.cs.uc4.authentication.impl.AuthenticationLoader
-    filters.enabled += "play.filters.gzip.GzipFilter"
-}
+play.application.loader = de.upb.cs.uc4.authentication.impl.AuthenticationLoader
 
 uc4.authentication {
     # the amount of days the refresh token is valid
