+++ product_code/authentication_service/impl/src/main/scala/de/upb/cs/uc4/authentication/impl/AuthenticationApplication.scala
@@ -8,41 +8,33 @@ import com.lightbend.lagom.scaladsl.broker.kafka.LagomKafkaComponents
 import com.lightbend.lagom.scaladsl.persistence.jdbc.JdbcPersistenceComponents
 import com.lightbend.lagom.scaladsl.persistence.slick.SlickPersistenceComponents
 import com.lightbend.lagom.scaladsl.playjson.JsonSerializerRegistry
-import com.lightbend.lagom.scaladsl.server.{ LagomApplication, LagomApplicationContext, LagomServer }
+import com.lightbend.lagom.scaladsl.server.{ LagomApplicationContext, LagomServer }
 import com.softwaremill.macwire.wire
 import de.upb.cs.uc4.authentication.api.AuthenticationService
 import de.upb.cs.uc4.authentication.impl.actor.{ AuthenticationBehaviour, AuthenticationState }
 import de.upb.cs.uc4.authentication.impl.commands.DeleteAuthentication
 import de.upb.cs.uc4.authentication.impl.readside.{ AuthenticationDatabase, AuthenticationEventProcessor }
 import de.upb.cs.uc4.authentication.model.JsonUsername
-import de.upb.cs.uc4.shared.server.Hashing
 import de.upb.cs.uc4.shared.server.messages.Confirmation
+import de.upb.cs.uc4.shared.server.{ Hashing, UC4Application }
 import de.upb.cs.uc4.user.api.UserService
 import play.api.db.HikariCPComponents
-import play.api.libs.ws.ahc.AhcWSComponents
-import play.api.mvc.EssentialFilter
-import play.filters.cors.CORSComponents
 
 import scala.concurrent.Future
 import scala.concurrent.duration._
 
 abstract class AuthenticationApplication(context: LagomApplicationContext)
-  extends LagomApplication(context)
+  extends UC4Application(context)
   with SlickPersistenceComponents
   with JdbcPersistenceComponents
   with HikariCPComponents
-  with LagomKafkaComponents
-  with CORSComponents
-  with AhcWSComponents {
+  with LagomKafkaComponents {
 
   private implicit val timeout: Timeout = Timeout(5.seconds)
 
   // Create ReadSide
   lazy val database: AuthenticationDatabase = wire[AuthenticationDatabase]
   lazy val processor: AuthenticationEventProcessor = wire[AuthenticationEventProcessor]
-
-  // Set HttpFilter to the default CorsFilter
-  override val httpFilters: Seq[EssentialFilter] = Seq(corsFilter)
 
   // Bind the service that this server provides
   override lazy val lagomServer: LagomServer = serverFor[AuthenticationService](wire[AuthenticationServiceImpl])
