+++ product_code/course_service/impl/src/main/resources/application.conf
@@ -1,9 +1,6 @@
 #
 #
-play {
-    application.loader = de.upb.cs.uc4.course.impl.CourseLoader
-    filters.enabled += "play.filters.gzip.GzipFilter"
-}
+play.application.loader = de.upb.cs.uc4.course.impl.CourseLoader
 
 db.default {
   driver = "org.postgresql.Driver"
