+++ product_code/matriculation_service/impl/src/main/resources/application.conf
@@ -1,9 +1,6 @@
 #
 #
-play {
-    application.loader = de.upb.cs.uc4.matriculation.impl.MatriculationLoader
-    filters.enabled += "play.filters.gzip.GzipFilter"
-}
+play.application.loader = de.upb.cs.uc4.matriculation.impl.MatriculationLoader
 
 uc4.hyperledger {
   username = cli
