+++ product_code/matriculation_service/impl/src/main/scala/de/upb/cs/uc4/matriculation/impl/MatriculationApplication.scala
@@ -1,28 +1,19 @@
 package de.upb.cs.uc4.matriculation.impl
 
 import com.lightbend.lagom.scaladsl.playjson.JsonSerializerRegistry
-import com.lightbend.lagom.scaladsl.server.{ LagomApplication, LagomApplicationContext, LagomServer }
+import com.lightbend.lagom.scaladsl.server.{ LagomApplicationContext, LagomServer }
 import com.softwaremill.macwire.wire
 import de.upb.cs.uc4.hyperledger.HyperledgerComponent
 import de.upb.cs.uc4.matriculation.api.MatriculationService
 import de.upb.cs.uc4.matriculation.impl.actor.MatriculationBehaviour
-import de.upb.cs.uc4.shared.server.AuthenticationComponent
+import de.upb.cs.uc4.shared.server.UC4Application
 import de.upb.cs.uc4.user.api.UserService
-import play.api.libs.ws.ahc.AhcWSComponents
-import play.api.mvc.EssentialFilter
-import play.filters.cors.CORSComponents
 
 abstract class MatriculationApplication(context: LagomApplicationContext)
-  extends LagomApplication(context)
-  with CORSComponents
-  with AhcWSComponents
-  with AuthenticationComponent
+  extends UC4Application(context)
   with HyperledgerComponent {
 
   override def createActorFactory: MatriculationBehaviour = wire[MatriculationBehaviour]
-
-  // Set HttpFilter to the default CorsFilter
-  override val httpFilters: Seq[EssentialFilter] = Seq(corsFilter)
 
   // Bind UserService
   lazy val userService: UserService = serviceClient.implement[UserService]
