+++ product_code/user_service/impl/src/main/resources/application.conf
@@ -1,10 +1,6 @@
 #
 #
-play {
-    application.loader = de.upb.cs.uc4.user.impl.UserLoader
-    filters.enabled += play.filters.gzip.GzipFilter
-    filters.gzip.threshold = 0
-}
+play.application.loader = de.upb.cs.uc4.user.impl.UserLoader
 
 db.default {
   driver = "org.postgresql.Driver"
